+++ backend/src/main/resources/application.properties
@@ -17,7 +17,6 @@ springdoc.swagger-ui.tags-sorter=alpha
 springdoc.swagger-ui.operations-sorter=alpha
 springdoc.api-docs.path=/api-docs
 
-
 springdoc.default-consumes-media-type=application/json;charset=UTF-8
 springdoc.default-produces-media-type=application/json;charset=UTF-8
